Remove the configuration class's dependency on ulabel
