+++ src/index.js
@@ -226,6 +226,49 @@ export class ULabel {
         return ret;
     }
 
+    static create_toolbox(ulabel, toolbox_item_order = null) {
+
+        //grab the default toolbox if one wasn't provided
+        if (toolbox_item_order == null) {
+            toolbox_item_order = ulabel.configuration.default_toolbox_item_order
+        }
+
+        //There's no point to having an empty toolbox, so throw an error if the toolbox is empty.
+        //The toolbox won't actually break if there aren't any items in the toolbox, so if for
+        //whatever reason we want that in the future, then feel free to remove this error.
+        if (toolbox_item_order.length == 0) {
+            throw new Error("No Toolbox Items Given")
+        }
+
+        let toolbox_instance_list = [];
+        //Go through the items in toolbox_item_order and add their instance to the toolbox instance list
+        for (let i = 0; i < toolbox_item_order.length; i++) {
+
+            let args, toolbox_key;
+
+            //If the value of toolbox_item_order[i] is a number then that means the it is one of the 
+            //enumerated toolbox items, so set it to the key, otherwise the element must be an array
+            //of which the first element of that array must be the enumerated value, and the arguments
+            //must be the second value
+            if (typeof(toolbox_item_order[i]) == "number") {
+                toolbox_key = toolbox_item_order[i]
+            } else {
+
+                toolbox_key = toolbox_item_order[i][0];
+                args = toolbox_item_order[i][1]  
+            }
+
+            let toolbox_item_class = ulabel.configuration.toolbox_map.get(toolbox_key);
+
+            if (args == null) {
+                toolbox_instance_list.push(new toolbox_item_class(ulabel))
+            } else {
+                toolbox_instance_list.push(new toolbox_item_class(ulabel, args))
+            }           
+        }
+
+        return toolbox_instance_list
+    }
 
     static prep_window_html(ul) {
         // Bring image and annotation scaffolding in
@@ -234,49 +277,11 @@ export class ULabel {
         // const tabs = ULabel.get_toolbox_tabs(ul);
         const images = ULabel.get_images_html(ul);
         const frame_annotation_dialogs = ULabel.get_frame_annotation_dialogs(ul);
-
-        //let configuration = new Configuration();
-
-        console.log(ul)
-
-        // let toolbox_item_order = ul.configuration.default_toolbox_item_order;
-
-        // //create the toolbox
-        // if (toolbox_item_order.length == 0) {
-        //     throw new Error("No Toolbox Items Given")
-        // }
-
-        // let toolbox_instance_list = [];
-        // //Go through the items in toolbox_item_order and add their instance to the toolbox instance list
-        // for (let i = 0; i < toolbox_item_order.length; i++) {
-
-        //     let args, toolbox_key;
-
-        //     //If the value of toolbox_item_order[i] is a number then that means the it is one of the 
-        //     //enumerated toolbox items, so set it to the key, otherwise the element must be an array
-        //     //of which the first element of that array must be the enumerated value, and the arguments
-        //     //must be the second value
-        //     if (typeof(toolbox_item_order[i]) == "number") {
-        //         toolbox_key = toolbox_item_order[i]
-        //     } else {
-
-        //         toolbox_key = toolbox_item_order[i][0];
-        //         args = toolbox_item_order[i][1]  
-        //     }
-
-        //     let toolbox_item_class = ul.Configuration.toolbox_map.get(toolbox_key);
-
-        //     if (args == null) {
-        //         toolbox_instance_list.push(new ul.Configuration.toolbox_item_class(ulabel))
-        //     } else {
-        //         toolbox_instance_list.push(new ul.Configuration.toolbox_item_class(ulabel, args))
-        //     }           
-        // }
         
         // const toolbox = configuration.create_toolbox();
         const toolbox = new Toolbox(
             [],
-            ul.configuration.create_toolbox(ul)
+            ULabel.create_toolbox(ul)
         );
